feat(app) Geocode all positions, then filter, then dispatch.
The geocoder receives all positions. It's up to itself to filter what
position is a candidate to be geocoded. For instance, position with no
name cannot be geocoded. So the geocoder returns `List (Maybe
LongitudeLatitude)` instead of `List (Int, LongitudeLatitude)` (the
`Int` was the index of the position, but since all of them are
received, no need to get the index). Then the rest of the modification
is about dealing with `Maybe`.

Finally, a line is drew between markers if there are more than 1 marker.